requirements.txt
opencv-python
moderngl
numpy
mediapipe
Pillow